Warning from a non-release firefox build
