+++ src/components/shared/Warning.js
@@ -5,12 +5,8 @@
 
 import React, { PureComponent } from 'react';
 import './Warning.css';
-import explicitConnect from '../../utils/connect';
-import { getProfile } from '../../selectors/profile';
-import type { Profile } from 'firefox-profiler/types';
-import type { ConnectedProps } from '../../utils/connect';
 
-type OwnProps = {|
+type Props = {|
   +message: string,
   +actionText?: string,
   +actionTitle?: string,
@@ -18,16 +14,11 @@ type OwnProps = {|
   +onClose?: () => mixed,
 |};
 
-type StateProps = {|
-  +profile: Profile,
-|};
-
-type Props = ConnectedProps<OwnProps, StateProps>;
 type State = {|
   +isNoticeDisplayed: boolean,
 |};
 
-class Warning extends PureComponent<Props, State> {
+export class Warning extends PureComponent<Props, State> {
   state = { isNoticeDisplayed: true };
 
   _onHideClick = () => {
@@ -45,51 +36,32 @@ class Warning extends PureComponent<Props, State> {
       return null;
     }
 
-    const {
-      message,
-      actionText,
-      actionTitle,
-      actionOnClick,
-      profile,
-    } = this.props;
-
-    const { meta } = profile;
+    const { message, actionText, actionTitle, actionOnClick } = this.props;
 
     return (
-      <>
-        {meta.debug ? (
-          <div className="warningMessageBarWrapper">
-            <div className="photon-message-bar photon-message-bar-warning warningMessageBar">
-              {message}
-              {actionText ? (
-                <button
-                  className="photon-button photon-button-micro photon-message-bar-action-button"
-                  type="button"
-                  title={actionTitle}
-                  aria-label={actionTitle}
-                  onClick={actionOnClick}
-                >
-                  {actionText}
-                </button>
-              ) : null}
-              <button
-                className="photon-button photon-message-bar-close-button"
-                type="button"
-                aria-label="Hide the message"
-                title="Hide the message"
-                onClick={this._onHideClick}
-              />
-            </div>
-          </div>
-        ) : null}
-      </>
+      <div className="warningMessageBarWrapper">
+        <div className="photon-message-bar photon-message-bar-warning warningMessageBar">
+          {message}
+          {actionText ? (
+            <button
+              className="photon-button photon-button-micro photon-message-bar-action-button"
+              type="button"
+              title={actionTitle}
+              aria-label={actionTitle}
+              onClick={actionOnClick}
+            >
+              {actionText}
+            </button>
+          ) : null}
+          <button
+            className="photon-button photon-message-bar-close-button"
+            type="button"
+            aria-label="Hide the message"
+            title="Hide the message"
+            onClick={this._onHideClick}
+          />
+        </div>
+      </div>
     );
   }
 }
-
-export default explicitConnect<StateProps>({
-  mapStateToProps: state => ({
-    profile: getProfile(state),
-  }),
-  component: Warning,
-});
